Spring Security Completed Order created

+++ src/main/java/telran/java57/farmmarket/dto/CreateProductDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @Builder
 public class CreateProductDto {
     private String name;
-    private int price;
+    private Double price;
     private int quantity;
     private String category;
 }
